Don't bomb out when close_to is called with no options.
The first iteration of the args, remaining loop sets self.args but if no
arguments are provided the assignment to self.args never takes place and
the first attempt to access self.args raises AttributeError.

Force a (fruitless) call to parse_args() so we instead fall through to
the usage message.