Support Coriolis ship dumps.
Read FSD information, along with any engineer modifications, from a
Coriolis JSON export.  The fuel_usage and edts commands can be told to
read the ship data via the --ship argument, and the FSD class, mass,
fuel tank and cargo capacity can be overridden as usual.
For edi we add the load_ship command to read the JSON file.