Handle passenger tours.
Passenger tours impose partial ordering constraints.  Suppose a
passenger wants to go to points A, B and C while a second passenger
wants to go to points X, Y and Z.  It is necessary to visit point A
before point B and point X before point Y.  However there are no
constraints regarding whether A or X are visited first.

A, B, X, Y is a valid route, as is A, X, Y, B and X, Y, A, B.

With the new -O flag, which may be given multiple times, we specify that
the following destinations form a passenger tour and must be visited in
order.  Any destinations given before the first use of -O will be
considered unordered.  Any given after a subsequent use of -O will form
a new tour.

The resultant route will satisfy all tour constraints in addition to any
other requirements.