ColumnObjectWriter class.
Use it to format rows nicely with aligned columns.

Instantiate the class:

  cow = ColumnObjectWriter()

Optionally provide the number of columns.

  cow = ColumnObjectWriter(3)

Optionally provide a list of alignments '<' or '>'.

  cow = ColumnObjectWriter(3, ['<', '>', '<'])

The first and third columns will be left justified.  The second will be
right justified.

Add rows.

  col.add(['row 1 column 1', 'row 1 column 2', 'row 1 column 3'])

Print it!

  col.out()

Or iterate the rows:

  for row in col.render():
    print(row)

Or convert the whole table to_string().