Differentiate between scale and suffix.
When instantiating a Dist the second argument is now the scale.  So for
instance Dist(1, 'Mm') will be created as 1,000,000m.

The optional third argument is the preferred suffix to be used when
prettyprinting, and may be different.

The new convert() method prints the distance in an arbitrary scale, and
can be called regardless of how the object was instantiated.