Handle ship reserve fuel tank.
All ships have a reserve fuel tank which is drained by powered modules
during flight.  The size of the tank varies between ships.

When refueling by 10% at a station the reserve tank is always filled,
meaning that a 10% refuel of a ship with a main tank of size n tonnes
and reserve tank of r tonnes actually fills (n + r) / 10 tonnes of fuel.

If the new --reserve-tank flag is passed to set_ship we now remember the
reserve tank size for later use when calculating a 10% refuel.