Allow specifying explicit refuels in fuel_usage.
Specify a refuel amount at any point in the argument list.  The amount
may be an explicit number of tonnes or a percentage of the main tank
size.  If a reserve tank size has been specified the percentage refuel
will take it into account.

Alternatively the refuel may be preceded by an = symbol to specify
setting the fuel to exactly that amount or percentage.