Scoop and station fuel strategies.
With both the "scoop" and "station" fuel strategies we walk the
candidate route in reverse and validate that we can arrive at each leg
with enough (and not too much) fuel to continue.

The "station" strategy allows refueling by some multiple of 10% at the
start of each hop.  Note that we currently are not able to verify that
the start of the hop actually is a station with refuel capability.

The "scoop" strategy additionally allows refueling at scoopable stars
if arrival star data is present in the database.

Note that the fuel strategy only validates that it is possible to
traverse the route.  It does not currently advise when and how much to
refuel.

A limitation of the initial implementation is that each hop is treated
independently.  The amount of fuel left over at the end of one hop is
not taken into consideration when plotting a subsequent hop.
Consequently there may be a better route than we can find.