Fix namespace when parsing close_to args.
Invocations of the form "close_to -m 10 Sol -m 20 Sirius" were not
working since the argument name was changed from 'system' to 'systems',
which was done to make the backend parameter name more meaningful.

Technically it only ever worked by coincidence that the argument name
was the same as the namespace.