v3 API returns results in JSON format.
A POST request to /api/v3/<app> will emit a JSON representation of the
opaque types returned by the backend Appliocations.

The output will contain two values, one with key "error" and one with
key "result".

If the backend throws an exception, the "error" will describe the error
and the "result" will be empty.

On success the "error" will be null and the "result" will be an array of
opaque entries.