Allow returning a direction as a bearing.
With the --bearing flag the direction will be returned as a bearing and
elevation measured in mils (6400 mils in a circle with 0 at north) or
with the --degrees flag as degrees.

The bearing and elevation are measured such that Sol to Sagittarius A*
is far north, slightly east and slightly down.

The --degrees flag is now also respected by --angle to return the vector
dot product in degrees.